+++ js/react-runner/package.json
@@ -2,6 +2,9 @@
   "name": "reach-react-app",
   "version": "0.0.0",
   "private": true,
+  "devDependencies": {
+    "crypto-browserify": "^3.12.0"
+  },
   "dependencies": {
     "@perawallet/connect": "^0.2.0",
     "algosdk": "^1.19.0",
